+++ noodles-vcf/src/io/reader/record_buf/filters.rs
@@ -7,8 +7,6 @@ use crate::variant::record_buf::Filters;
 pub enum ParseError {
     /// The input is empty.
     Empty,
-    /// A filter is invalid.
-    InvalidFilter,
     /// A filter is duplicated.
     DuplicateFilter,
 }
@@ -19,7 +17,6 @@ impl fmt::Display for ParseError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             Self::Empty => write!(f, "empty input"),
-            Self::InvalidFilter => write!(f, "invalid filter"),
             Self::DuplicateFilter => write!(f, "duplicate filter"),
         }
     }
@@ -42,19 +39,10 @@ pub(super) fn parse_filters(s: &str, filters: &mut Filters) -> Result<(), ParseE
     for raw_filter in s.split(DELIMITER) {
         if !filters.insert(raw_filter.into()) {
             return Err(ParseError::DuplicateFilter);
-        } else if !is_valid_filter(raw_filter) {
-            return Err(ParseError::InvalidFilter);
         }
     }
 
     Ok(())
-}
-
-fn is_valid_filter(s: &str) -> bool {
-    match s {
-        "" | "0" => false,
-        _ => s.chars().all(|c| !c.is_whitespace()),
-    }
 }
 
 #[cfg(test)]
@@ -80,26 +68,7 @@ mod tests {
         );
 
         assert_eq!(parse_filters("", &mut filters), Err(ParseError::Empty));
-        assert_eq!(
-            parse_filters("0", &mut filters),
-            Err(ParseError::InvalidFilter)
-        );
-        assert_eq!(
-            parse_filters("q 10", &mut filters),
-            Err(ParseError::InvalidFilter)
-        );
-        assert_eq!(
-            parse_filters(";q10", &mut filters),
-            Err(ParseError::InvalidFilter)
-        );
-        assert_eq!(
-            parse_filters("q10;;s50", &mut filters),
-            Err(ParseError::InvalidFilter)
-        );
-        assert_eq!(
-            parse_filters("q10;", &mut filters),
-            Err(ParseError::InvalidFilter)
-        );
+
         assert_eq!(
             parse_filters("q10;q10", &mut filters),
             Err(ParseError::DuplicateFilter)